+++ zcash_client_backend/src/data_api/testing.rs
@@ -85,11 +85,11 @@ use {
 };
 
 use super::{
-    Account, AccountBalance, AccountBirthday, AccountMeta, AccountPurpose, AccountSource,
-    AddressInfo, BlockMetadata, DecryptedTransaction, InputSource, NoteFilter, NullifierQuery,
-    ReceivedNotes, ReceivedTransactionOutput, SAPLING_SHARD_HEIGHT, ScannedBlock, SeedRelevance,
-    SentTransaction, TransactionDataRequest, TransactionStatus, WalletCommitmentTrees, WalletRead,
-    WalletSummary, WalletTest, WalletWrite, Zip32Derivation,
+    Account, AccountBalance, AccountBirthday, AccountMeta, AccountSource, AddressInfo,
+    BlockMetadata, DecryptedTransaction, InputSource, NoteFilter, NullifierQuery, ReceivedNotes,
+    ReceivedTransactionOutput, SAPLING_SHARD_HEIGHT, ScannedBlock, SeedRelevance, SentTransaction,
+    TransactionDataRequest, TransactionStatus, WalletCommitmentTrees, WalletRead, WalletSummary,
+    WalletTest, WalletWrite, Zip32Derivation,
     anchor_retention::AnchorRetentionInterval,
     chain::{BlockSource, ChainState, CommitmentTreeRoot, ScanSummary, scan_cached_blocks},
     error::Error,
@@ -3578,7 +3578,7 @@ impl WalletWrite for MockWalletDb {
         _account_name: &str,
         _unified_key: &UnifiedFullViewingKey,
         _birthday: &AccountBirthday,
-        _purpose: AccountPurpose,
+        _derivation: Option<Zip32Derivation>,
         _key_source: Option<&str>,
     ) -> Result<Self::Account, <Self as WalletRead>::Error> {
         todo!()