+++ zcash_client_sqlite/src/wallet.rs
@@ -103,10 +103,10 @@ use zcash_address::ZcashAddress;
 use zcash_client_backend::{
     DecryptedOutput,
     data_api::{
-        Account as _, AccountBalance, AccountBirthday, AccountPurpose, AccountSource, AddressInfo,
-        AddressSource, BlockMetadata, Progress, Ratio, ReceivedTransactionOutput,
-        SAPLING_SHARD_HEIGHT, SentTransaction, SentTransactionOutput, TransactionDataRequest,
-        TransactionStatus, WalletSummary, Zip32Derivation,
+        Account as _, AccountBalance, AccountBirthday, AccountSource, AddressInfo, AddressSource,
+        BlockMetadata, Progress, Ratio, ReceivedTransactionOutput, SAPLING_SHARD_HEIGHT,
+        SentTransaction, SentTransactionOutput, TransactionDataRequest, TransactionStatus,
+        WalletSummary, Zip32Derivation,
         anchor_retention::AnchorRetentionInterval,
         chain::ChainState,
         defaults::address_receiver_matches_ua,
@@ -210,7 +210,6 @@ fn parse_account_source(
     hd_seed_fingerprint: Option<[u8; 32]>,
     hd_account_index: Option<u32>,
     #[cfg(feature = "zcashd-compat")] legacy_account_index: i64,
-    spending_key_available: bool,
     key_source: Option<String>,
 ) -> Result<AccountSource, SqliteClientError> {
     let derivation = hd_seed_fingerprint
@@ -240,11 +239,7 @@ fn parse_account_source(
             key_source,
         }),
         (1, derivation) => Ok(AccountSource::Imported {
-            purpose: if spending_key_available {
-                AccountPurpose::Spending { derivation }
-            } else {
-                AccountPurpose::ViewOnly
-            },
+            derivation,
             key_source,
         }),
         (0, None) => Err(SqliteClientError::CorruptedData(
@@ -489,19 +484,15 @@ pub(crate) fn add_account<P: consensus::Parameters>(
 
     let account_uuid = AccountUuid(Uuid::new_v4());
 
-    let (derivation, spending_key_available, key_source) = match kind {
+    let (derivation, key_source) = match kind {
         AccountSource::Derived {
             derivation,
             key_source,
-        } => (Some(derivation), true, key_source),
+        } => (Some(derivation), key_source),
         AccountSource::Imported {
-            purpose: AccountPurpose::Spending { derivation },
+            derivation,
             key_source,
-        } => (derivation.as_ref(), true, key_source),
-        AccountSource::Imported {
-            purpose: AccountPurpose::ViewOnly,
-            key_source,
-        } => (None, false, key_source),
+        } => (derivation.as_ref(), key_source),
     };
 
     let ivk_cache = IvkItemCache::from_uivk(&uivk);
@@ -531,8 +522,7 @@ pub(crate) fn add_account<P: consensus::Parameters>(
                 ufvk, uivk,
                 orchard_ivk_item_cache, sapling_ivk_item_cache, p2pkh_ivk_item_cache,
                 birthday_height, birthday_sapling_tree_size, birthday_orchard_tree_size,
-                recover_until_height,
-                has_spend_key
+                recover_until_height
             )
             VALUES (
                 :account_name,
@@ -543,8 +533,7 @@ pub(crate) fn add_account<P: consensus::Parameters>(
                 :ufvk, :uivk,
                 :orchard_ivk_item_cache, :sapling_ivk_item_cache, :p2pkh_ivk_item_cache,
                 :birthday_height, :birthday_sapling_tree_size, :birthday_orchard_tree_size,
-                :recover_until_height,
-                :has_spend_key
+                :recover_until_height
             )
             RETURNING id
             "#,
@@ -565,7 +554,6 @@ pub(crate) fn add_account<P: consensus::Parameters>(
                 ":birthday_sapling_tree_size": birthday_sapling_tree_size,
                 ":birthday_orchard_tree_size": birthday_orchard_tree_size,
                 ":recover_until_height": birthday.recover_until().map(u32::from),
-                ":has_spend_key": i64::from(spending_key_available),
             ],
             |row| row.get(0).map(AccountRef),
         )
@@ -1802,7 +1790,6 @@ fn parse_account_row<P: consensus::Parameters>(
         row.get("hd_account_index")?,
         #[cfg(feature = "zcashd-compat")]
         row.get("zcashd_legacy_address_index")?,
-        row.get("has_spend_key")?,
         row.get("key_source")?,
     )?;
 
@@ -1849,7 +1836,7 @@ pub(crate) fn get_account<P: Parameters>(
         r#"
         SELECT id, name, uuid, account_kind,
                hd_seed_fingerprint, hd_account_index, zcashd_legacy_address_index, key_source,
-               ufvk, uivk, has_spend_key, birthday_height
+               ufvk, uivk, birthday_height
         FROM accounts
         WHERE uuid = :account_uuid
         "#,
@@ -1873,7 +1860,7 @@ pub(crate) fn get_account_internal<P: Parameters>(
         r#"
         SELECT id, name, uuid, account_kind,
                hd_seed_fingerprint, hd_account_index, zcashd_legacy_address_index, key_source,
-               ufvk, uivk, has_spend_key, birthday_height
+               ufvk, uivk, birthday_height
         FROM accounts
         WHERE id = :account_id
         "#,
@@ -1910,7 +1897,7 @@ pub(crate) fn get_account_for_uivk<P: consensus::Parameters>(
     let mut stmt = conn.prepare(
         "SELECT id, name, uuid, account_kind,
                 hd_seed_fingerprint, hd_account_index, zcashd_legacy_address_index, key_source,
-                ufvk, uivk, has_spend_key, birthday_height
+                ufvk, uivk, birthday_height
          FROM accounts
          WHERE orchard_ivk_item_cache = :orchard_ivk_item_cache
             OR sapling_ivk_item_cache = :sapling_ivk_item_cache
@@ -1992,7 +1979,7 @@ fn upgrade_account_ufvk<P: consensus::Parameters>(
     let mut stmt = conn.prepare_cached(
         "SELECT id, name, uuid, account_kind,
                 hd_seed_fingerprint, hd_account_index, zcashd_legacy_address_index, key_source,
-                ufvk, uivk, has_spend_key, birthday_height
+                ufvk, uivk, birthday_height
          FROM accounts
          WHERE id = :account_id",
     )?;
@@ -2050,7 +2037,7 @@ fn upgrade_account_uivk<P: consensus::Parameters>(
     let mut stmt = conn.prepare_cached(
         "SELECT id, name, uuid, account_kind,
                 hd_seed_fingerprint, hd_account_index, zcashd_legacy_address_index, key_source,
-                ufvk, uivk, has_spend_key, birthday_height
+                ufvk, uivk, birthday_height
          FROM accounts
          WHERE id = :account_id",
     )?;